+++ CHANGELOG.md
@@ -5,8 +5,11 @@
         - Plugins supported: `table`
     - API:
       - Support for exits numbers and names. New member `exits` in `RouteStep`, based on `junction:ref` on ways
+      - `RouteStep` now has new parameter `classes` that can be set in the profile on each way.
     - Profiles:
         - `result.exits` allows you to set a way's exit numbers and names, see [`junction:ref`](http://wiki.openstreetmap.org/wiki/Proposed_features/junction_details)
+        - `ExtractionWay` now as new property `forward_classes` and `backward_classes` that can set in the `way_function`.
+           The maximum number of classes is 8.
 
 # 5.8.0
   - Changes from 5.7